Optimize imports and fix refactor errors
