+++ data/src/test/java/org/oppia/android/data/backends/gae/api/ClassroomServiceTest.kt
@@ -1,4 +1,4 @@
-package org.oppia.android.data.backends.test
+package org.oppia.android.data.backends.gae.api
 
 import android.app.Application
 import androidx.test.core.app.ApplicationProvider
@@ -9,7 +9,6 @@ import dagger.Component
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.oppia.android.data.backends.gae.api.ClassroomService
 import org.oppia.android.testing.network.MockClassroomService
 import org.oppia.android.testing.network.RetrofitTestModule
 import org.robolectric.annotation.LooperMode
@@ -45,7 +44,7 @@ class ClassroomServiceTest {
   }
 
   private fun setUpTestApplicationComponent() {
-    DaggerMockClassroomTest_TestApplicationComponent
+    DaggerClassroomServiceTest_TestApplicationComponent
       .builder()
       .setApplication(ApplicationProvider.getApplicationContext()).build().inject(this)
   }
@@ -62,6 +61,6 @@ class ClassroomServiceTest {
       fun build(): TestApplicationComponent
     }
 
-    fun inject(test: MockClassroomTest)
+    fun inject(test: ClassroomServiceTest)
   }
 }
