+++ data/src/test/java/org/oppia/android/data/backends/gae/api/StoryServiceTest.kt
@@ -1,11 +1,16 @@
 package org.oppia.android.data.backends.gae.api
 
+import android.app.Application
+import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth.assertThat
+import dagger.BindsInstance
+import dagger.Component
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.oppia.android.data.backends.api.MockStoryService
+import org.oppia.android.testing.network.MockStoryService
+import org.oppia.android.testing.network.RetrofitTestModule
 import org.robolectric.annotation.LooperMode
 import retrofit2.mock.MockRetrofit
 import javax.inject.Inject
@@ -39,7 +44,7 @@ class StoryServiceTest {
   }
 
   private fun setUpTestApplicationComponent() {
-    DaggerMockStoryTest_TestApplicationComponent
+    DaggerStoryServiceTest_TestApplicationComponent
       .builder()
       .setApplication(ApplicationProvider.getApplicationContext()).build().inject(this)
   }
@@ -56,6 +61,6 @@ class StoryServiceTest {
       fun build(): TestApplicationComponent
     }
 
-    fun inject(test: MockStoryTest)
+    fun inject(test: StoryServiceTest)
   }
 }
