+++ data/src/test/java/org/oppia/android/data/backends/gae/api/TopicServiceTest.kt
@@ -1,12 +1,17 @@
 package org.oppia.android.data.backends.test
 
+import android.app.Application
+import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth.assertThat
+import dagger.BindsInstance
+import dagger.Component
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.oppia.android.data.backends.api.MockTopicService
 import org.oppia.android.data.backends.gae.api.TopicService
+import org.oppia.android.testing.network.MockTopicService
+import org.oppia.android.testing.network.RetrofitTestModule
 import org.robolectric.annotation.LooperMode
 import retrofit2.mock.MockRetrofit
 import javax.inject.Inject
@@ -40,7 +45,7 @@ class TopicServiceTest {
   }
 
   private fun setUpTestApplicationComponent() {
-    DaggerMockTopicTest_TestApplicationComponent
+    DaggerTopicServiceTest_TestApplicationComponent
       .builder()
       .setApplication(ApplicationProvider.getApplicationContext()).build().inject(this)
   }
@@ -57,6 +62,6 @@ class TopicServiceTest {
       fun build(): TestApplicationComponent
     }
 
-    fun inject(test: MockTopicTest)
+    fun inject(test: TopicServiceTest)
   }
 }
